CORE-7356 Refactor DiskResourcesMovedEvent.
Moved DiskResourcesMovedEvent into
org.iplantc.de.client.events.diskResources so that it can be fired from
the DiskResourceServiceFacadeImpl, instead of from the
DiskResourceMoveCallback.
Updated
org.iplantc.de.client.services.impl.DiskResourceServiceFacadeImpl#refreshFolder
to fire its FolderRefreshedEvent before calling its callback's onSuccess
method. This allows data window nav-panels to handle the refreshed
folders before attempting to re-select folders after the move.
